+++ test/unit/ui/app/reducers/metamask.spec.js
@@ -309,6 +309,8 @@ describe('MetaMask Reducers', () => {
       errors: {},
       editingTransactionId: 22,
       forceGasMin: '0xGas',
+      ensResolution: null,
+      ensResolutionError: '',
     }
 
     const sendState = reduceMetamask({}, {
